Fix 구조
header 생성
