+++ RhythmGame/RhythmGame/Release/RhythmGame.hex
@@ -1,16 +1,16 @@
 :1000000045C0000060C000005EC000005CC0000091
 :100010005AC0000058C0000056C0000054C0000084
 :1000200052C0000050C000004EC000004CC0000094
-:100030004AC0000048C000008DC1000044C000005C
+:100030004AC0000048C0000079C1000044C0000070
 :1000400042C0000040C000003EC000003CC00000B4
 :100050003AC0000038C0000036C0000034C00000C4
 :1000600032C0000030C000002EC000002CC00000D4
 :100070002AC0000028C0000026C0000024C00000E4
 :1000800022C0000020C000001EC0000011241FBEBE
-:10009000CFEFD0E1DEBFCDBF12E0A0E0B1E0E0ECF9
+:10009000CFEFD0E1DEBFCDBF12E0A0E0B1E0E0E7FE
 :1000A000F3E000E00BBF02C007900D92A03FB10744
-:1000B000D9F722E0A0EFB2E001C01D92A43FB20741
-:1000C000E1F704D17BC19CCF1D9A95E39A95F1F796
+:1000B000D9F722E0A0EFB2E001C01D92A23FB20743
+:1000C000E1F704D153C19CCF1D9A95E39A95F1F7BE
 :1000D00000001C981E981F9A8BBB85E38A95F1F748
 :1000E00000001D9808951D9A95E39A95F1F7000078
 :1000F0001C981E9A1F988BBB85E38A95F1F7000028
@@ -46,47 +46,42 @@
 :1002D00088BB87BB11BA82B9109265008093640015
 :1002E0008FE310DF88EB0EDF80E40CDFDDDF4BED0A
 :1002F00052E060E080E079DF8BE08DBD8CED8CBD5D
-:100300001FBC84E08EBD87BF78941092F3021092D8
-:10031000F0021092F1021092F2024091F302405D5D
-:1003200060E081E02BDF4091F002405D61E081E020
-:1003300025DF4AE362E081E021DF4091F102405D88
-:1003400063E081E01BDF4091F202405D64E081E008
-:1003500015DFE3CF1F920F920FB60F9211248F93E8
-:100360008091F2028F5F8A3019F08093F20220C0F0
-:100370001092F2028091F1028F5F863019F0809323
-:10038000F10216C01092F1028091F0028F5F8A3064
-:1003900019F08093F0020CC01092F0028091F302E9
-:1003A0008F5F863019F08093F30202C01092F3023F
-:1003B0008F910F900FBE0F901F901895F894FFCF5C
-:1003C000000000000000004F0000000700070014BC
-:1003D0007F147F14242A7F2A12231308646236496B
-:1003E0005522500005030000001C2241000041225C
-:1003F0001C0014083E081408083E0808005030008D
-:1004000000080808080800606000002010080402C6
-:100410003E5149453E00427F4000426151494621DC
-:1004200041454B311814127F1027454545393C4A48
-:100430004949300171090503364949493606494998
-:10044000291E0036360000005636000008142241EE
-:100450000014141414140041221408020151090656
-:10046000324979413E7E1111117E7F494949363E1C
-:10047000414141227F4141221C7F494949417F0935
-:100480000909013E4149497A7F0808087F00417FF8
-:1004900041002040413F017F081422417F404040FD
-:1004A000407F020C027F7F0408107F3E4141413EA5
-:1004B0007F090909063E4151215E7F091929462617
-:1004C0004949493201017F01013F4040403F1F201F
-:1004D00040201F3F4038403F6314081463070870F2
-:1004E00008076151494543007F414100020408105B
-:1004F000200041417F00040201020440404040408E
-:10050000000102040020545454787F484444383891
-:1005100044444420384444487F3854545418087E96
-:100520000901020C5252523E7F0804047800047DF7
-:1005300000002040443D007F1028440000417F40DF
-:10054000007C0418047C7C08040478384444443853
-:100550007C14141408081414187C7C08040408483B
-:1005600054545420043F4440203C4040207C1C20F4
-:1005700040201C3C4030403C44281028440C505043
-:10058000503C4464544C440008364100000077005D
-:100590000000413608000804081008436C6F636BC4
-:1005A00020496E746572727570742054657374009E
+:100300001FBC84E08EBD87BF78941092F0021092DB
+:10031000F1024091F002405D63E081E02FDF409107
+:10032000F102405D64E081E029DFF3CF1F920F927C
+:100330000FB60F9211248F938091F1028F5F8A3054
+:1003400019F08093F1020CC01092F1028091F0023A
+:100350008F5F863019F08093F00202C01092F00295
+:100360008F910F900FBE0F901F901895F894FFCFAC
+:10037000000000000000004F00000007000700140C
+:100380007F147F14242A7F2A1223130864623649BB
+:100390005522500005030000001C224100004122AC
+:1003A0001C0014083E081408083E080800503000DD
+:1003B0000008080808080060600000201008040217
+:1003C0003E5149453E00427F40004261514946212D
+:1003D00041454B311814127F1027454545393C4A99
+:1003E00049493001710905033649494936064949E9
+:1003F000291E00363600000056360000081422413F
+:1004000000141414141400412214080201510906A6
+:10041000324979413E7E1111117E7F494949363E6C
+:10042000414141227F4141221C7F494949417F0985
+:100430000909013E4149497A7F0808087F00417F48
+:1004400041002040413F017F081422417F4040404D
+:10045000407F020C027F7F0408107F3E4141413EF5
+:100460007F090909063E4151215E7F091929462667
+:100470004949493201017F01013F4040403F1F206F
+:1004800040201F3F4038403F631408146307087042
+:1004900008076151494543007F41410002040810AB
+:1004A000200041417F0004020102044040404040DE
+:1004B000000102040020545454787F4844443838E2
+:1004C00044444420384444487F3854545418087EE7
+:1004D0000901020C5252523E7F0804047800047D48
+:1004E00000002040443D007F1028440000417F4030
+:1004F000007C0418047C7C080404783844444438A4
+:100500007C14141408081414187C7C08040408488B
+:1005100054545420043F4440203C4040207C1C2044
+:1005200040201C3C4030403C44281028440C505093
+:10053000503C4464544C44000836410000007700AD
+:100540000000413608000804081008436C6F636B14
+:1005500020496E74657272757074205465737400EE
 :00000001FF
