+++ RhythmGame/RhythmGame/main.c
@@ -1,429 +1,48 @@
-//#include "lcd.h"
-#include <avr/io.h>
-#include <avr/interrupt.h>
-#define F_CPU 16000000UL
-#include <util/delay.h>
+#include "lcd.h"
 
+unsigned char sec0, sec1;
 
-#define SetBit(x,y)  (x|=(1<<y))
-#define ClrBit(x,y)  (x&=~(1<<y))
+int main(void)
+{
+	//Initialization
+	DDRA = 0xFF;
+	PORTB = 0xFF; DDRB = 0xFF;
+	DDRD = 0x0;
+	DDRE = 0xFF;
+	PORTG = 0x0; DDRG = 0xFF;
 
-
-unsigned char font[95][5] = {            /* 5x7 ASCII character font */
-	{0x00, 0x00, 0x00, 0x00, 0x00},    /* 0x20 space */
-	{0x00, 0x00, 0x4f, 0x00, 0x00},    /* 0x21 ! */
-	{0x00, 0x07, 0x00, 0x07, 0x00},    /* 0x22 " */
-	{0x14, 0x7f, 0x14, 0x7f, 0x14},    /* 0x23 # */
-	{0x24, 0x2a, 0x7f, 0x2a, 0x12},    /* 0x24 $ */
-	{0x23, 0x13, 0x08, 0x64, 0x62},    /* 0x25 % */
-	{0x36, 0x49, 0x55, 0x22, 0x50},    /* 0x26 & */
-	{0x00, 0x05, 0x03, 0x00, 0x00},    /* 0x27 ' */
-	{0x00, 0x1c, 0x22, 0x41, 0x00},    /* 0x28 ( */
-	{0x00, 0x41, 0x22, 0x1c, 0x00},    /* 0x29 ) */
-	{0x14, 0x08, 0x3e, 0x08, 0x14},    /* 0x2a * */
-	{0x08, 0x08, 0x3e, 0x08, 0x08},    /* 0x2b + */
-	{0x00, 0x50, 0x30, 0x00, 0x00},    /* 0x2c , */
-	{0x08, 0x08, 0x08, 0x08, 0x08},    /* 0x2d - */
-	{0x00, 0x60, 0x60, 0x00, 0x00},    /* 0x2e . */
-	{0x20, 0x10, 0x08, 0x04, 0x02},    /* 0x2f / */
-	{0x3e, 0x51, 0x49, 0x45, 0x3e},    /* 0x30 0 */
-	{0x00, 0x42, 0x7f, 0x40, 0x00},    /* 0x31 1 */
-	{0x42, 0x61, 0x51, 0x49, 0x46},    /* 0x32 2 */
-	{0x21, 0x41, 0x45, 0x4b, 0x31},    /* 0x33 3 */
-	{0x18, 0x14, 0x12, 0x7f, 0x10},    /* 0x34 4 */
-	{0x27, 0x45, 0x45, 0x45, 0x39},    /* 0x35 5 */
-	{0x3c, 0x4a, 0x49, 0x49, 0x30},    /* 0x36 6 */
-	{0x01, 0x71, 0x09, 0x05, 0x03},    /* 0x37 7 */
-	{0x36, 0x49, 0x49, 0x49, 0x36},    /* 0x38 8 */
-	{0x06, 0x49, 0x49, 0x29, 0x1e},    /* 0x39 9 */
-	{0x00, 0x36, 0x36, 0x00, 0x00},    /* 0x3a : */
-	{0x00, 0x56, 0x36, 0x00, 0x00},    /* 0x3b ; */
-	{0x08, 0x14, 0x22, 0x41, 0x00},    /* 0x3c < */
-	{0x14, 0x14, 0x14, 0x14, 0x14},    /* 0x3d = */
-	{0x00, 0x41, 0x22, 0x14, 0x08},    /* 0x3e > */
-	{0x02, 0x01, 0x51, 0x09, 0x06},    /* 0x3f ? */
-	{0x32, 0x49, 0x79, 0x41, 0x3e},    /* 0x40 @ */
-	{0x7e, 0x11, 0x11, 0x11, 0x7e},    /* 0x41 A */
-	{0x7f, 0x49, 0x49, 0x49, 0x36},    /* 0x42 B */
-	{0x3e, 0x41, 0x41, 0x41, 0x22},    /* 0x43 C */
-	{0x7f, 0x41, 0x41, 0x22, 0x1c},    /* 0x44 D */
-	{0x7f, 0x49, 0x49, 0x49, 0x41},    /* 0x45 E */
-	{0x7f, 0x09, 0x09, 0x09, 0x01},    /* 0x46 F */
-	{0x3e, 0x41, 0x49, 0x49, 0x7a},    /* 0x47 G */
-	{0x7f, 0x08, 0x08, 0x08, 0x7f},    /* 0x48 H */
-	{0x00, 0x41, 0x7f, 0x41, 0x00},    /* 0x49 I */
-	{0x20, 0x40, 0x41, 0x3f, 0x01},    /* 0x4a J */
-	{0x7f, 0x08, 0x14, 0x22, 0x41},    /* 0x4b K */
-	{0x7f, 0x40, 0x40, 0x40, 0x40},    /* 0x4c L */
-	{0x7f, 0x02, 0x0c, 0x02, 0x7f},    /* 0x4d M */
-	{0x7f, 0x04, 0x08, 0x10, 0x7f},    /* 0x4e N */
-	{0x3e, 0x41, 0x41, 0x41, 0x3e},    /* 0x4f O */
-	{0x7f, 0x09, 0x09, 0x09, 0x06},    /* 0x50 P */
-	{0x3e, 0x41, 0x51, 0x21, 0x5e},    /* 0x51 Q */
-	{0x7f, 0x09, 0x19, 0x29, 0x46},    /* 0x52 R */
-	{0x26, 0x49, 0x49, 0x49, 0x32},    /* 0x53 S */
-	{0x01, 0x01, 0x7f, 0x01, 0x01},    /* 0x54 T */
-	{0x3f, 0x40, 0x40, 0x40, 0x3f},    /* 0x55 U */
-	{0x1f, 0x20, 0x40, 0x20, 0x1f},    /* 0x56 V */
-	{0x3f, 0x40, 0x38, 0x40, 0x3f},    /* 0x57 W */
-	{0x63, 0x14, 0x08, 0x14, 0x63},    /* 0x58 X */
-	{0x07, 0x08, 0x70, 0x08, 0x07},    /* 0x59 Y */
-	{0x61, 0x51, 0x49, 0x45, 0x43},    /* 0x5a Z */
-	{0x00, 0x7f, 0x41, 0x41, 0x00},    /* 0x5b [ */
-	{0x02, 0x04, 0x08, 0x10, 0x20},    /* 0x5c \ */
-	{0x00, 0x41, 0x41, 0x7f, 0x00},    /* 0x5d ] */
-	{0x04, 0x02, 0x01, 0x02, 0x04},    /* 0x5e ^ */
-	{0x40, 0x40, 0x40, 0x40, 0x40},    /* 0x5f _ */
-	{0x00, 0x01, 0x02, 0x04, 0x00},    /* 0x60 ` */
-	{0x20, 0x54, 0x54, 0x54, 0x78},    /* 0x61 a */
-	{0x7f, 0x48, 0x44, 0x44, 0x38},    /* 0x62 b */
-	{0x38, 0x44, 0x44, 0x44, 0x20},    /* 0x63 c */
-	{0x38, 0x44, 0x44, 0x48, 0x7f},    /* 0x64 d */
-	{0x38, 0x54, 0x54, 0x54, 0x18},    /* 0x65 e */
-	{0x08, 0x7e, 0x09, 0x01, 0x02},    /* 0x66 f */
-	{0x0c, 0x52, 0x52, 0x52, 0x3e},    /* 0x67 g */
-	{0x7f, 0x08, 0x04, 0x04, 0x78},    /* 0x68 h */
-	{0x00, 0x04, 0x7d, 0x00, 0x00},    /* 0x69 i */
-	{0x20, 0x40, 0x44, 0x3d, 0x00},    /* 0x6a j */
-	{0x7f, 0x10, 0x28, 0x44, 0x00},    /* 0x6b k */
-	{0x00, 0x41, 0x7f, 0x40, 0x00},    /* 0x6c l */
-	{0x7c, 0x04, 0x18, 0x04, 0x7c},    /* 0x6d m */
-	{0x7c, 0x08, 0x04, 0x04, 0x78},    /* 0x6e n */
-	{0x38, 0x44, 0x44, 0x44, 0x38},    /* 0x6f o */
-	{0x7c, 0x14, 0x14, 0x14, 0x08},    /* 0x70 p */
-	{0x08, 0x14, 0x14, 0x18, 0x7c},    /* 0x71 q */
-	{0x7c, 0x08, 0x04, 0x04, 0x08},    /* 0x72 r */
-	{0x48, 0x54, 0x54, 0x54, 0x20},    /* 0x73 s */
-	{0x04, 0x3f, 0x44, 0x40, 0x20},    /* 0x74 t */
-	{0x3c, 0x40, 0x40, 0x20, 0x7c},    /* 0x75 u */
-	{0x1c, 0x20, 0x40, 0x20, 0x1c},    /* 0x76 v */
-	{0x3c, 0x40, 0x30, 0x40, 0x3c},    /* 0x77 w */
-	{0x44, 0x28, 0x10, 0x28, 0x44},    /* 0x78 x */
-	{0x0c, 0x50, 0x50, 0x50, 0x3c},    /* 0x79 y */
-	{0x44, 0x64, 0x54, 0x4c, 0x44},    /* 0x7a z */
-	{0x00, 0x08, 0x36, 0x41, 0x00},    /* 0x7b { */
-	{0x00, 0x00, 0x77, 0x00, 0x00},    /* 0x7c | */
-	{0x00, 0x41, 0x36, 0x08, 0x00},    /* 0x7d } */
-	{0x08, 0x04, 0x08, 0x10, 0x08}};   /* 0x7e ~ */
-
-
-	void cmd_l( unsigned char cmd )
-	{
-		SetBit(PORTE,5);	// E
-		_delay_us(10);
-		ClrBit(PORTE,4);	// RS
-		ClrBit(PORTE,6);	// CS2
-		SetBit(PORTE,7);	// CS1
-		PORTA = cmd ;
-		_delay_us(10);
-		ClrBit(PORTE,5);	// E
-	}
-
-	void cmd_r( unsigned char cmd )
-	{
-		SetBit(PORTE,5);	// E
-		_delay_us(10);
-		ClrBit(PORTE,4);	// RS
-		SetBit(PORTE,6);	// CS2
-		ClrBit(PORTE,7);	// CS1
-		PORTA = cmd ;
-		_delay_us(10);
-		ClrBit(PORTE,5);	// E
-	}
-
-	void cmd_lr( unsigned char cmd )
-	{
-		SetBit(PORTE,5);	// E
-		_delay_us(10);
-		ClrBit(PORTE,4);	// RS
-		SetBit(PORTE,6);	// CS2
-		SetBit(PORTE,7);	// CS1
-		PORTA = cmd ;
-		_delay_us(10);
-		ClrBit(PORTE,5);	// E
-	}
-
-	void data_l( unsigned char data )
-	{
-		SetBit(PORTE,5);	// E
-		_delay_us(10);
-		SetBit(PORTE,4);	// RS
-		ClrBit(PORTE,6);	// CS2
-		SetBit(PORTE,7);	// CS1
-		PORTA = data ;
-		_delay_us(10);
-		ClrBit(PORTE,5);	// E
-	}
-
-	void data_r( unsigned char data )
-	{
-		SetBit(PORTE,5);	// E
-		_delay_us(10);
-		SetBit(PORTE,4);	// RS
-		SetBit(PORTE,6);	// CS2
-		ClrBit(PORTE,7);	// CS1
-		PORTA = data ;
-		_delay_us(10);
-		ClrBit(PORTE,5);	// E
-	}
-
-	void data_lr( unsigned char data )
-	{
-		SetBit(PORTE,5);	// E
-		_delay_us(10);
-		SetBit(PORTE,4);	// RS
-		SetBit(PORTE,6);	// CS2
-		SetBit(PORTE,7);	// CS1
-		PORTA = data ;
-		_delay_us(10);
-		ClrBit(PORTE,5);	// E
-	}
-
-	void display_char(unsigned char x, unsigned char y, unsigned char character)
-	{
-		unsigned char i;
+	cmd_lr( 0x3F ); //DISP ON
+	cmd_lr( 0xB8 );
+	cmd_lr( 0x40 );
 		
-		cmd_lr(0xB8+x);     // x address
-		if( y < 10 )
-		cmd_l(0x40+y*6+4); // y address
-		else
-		cmd_r(0x40+(y-10)*6); // y address
+	lcd_clear();
+	display_string(0, 0, "Clock Interrupt Test");
 		
-		for(i = 0; i < 5; i++)
-		{
-			if(y < 10)
-			data_l(font[character-0x20][i]);
-			else
-			data_r(font[character-0x20][i]);
-		}
+	TCNT1H = 0xB;
+	TCNT1L = 0xDC;
+	TCCR1A = 0x0;
+	TCCR1B = 0x04;
 		
-		if(y < 10)
-		data_l(0x00);
-		else
-		data_r(0x00);
+	TIMSK = (1<<TOIE1);
+	sei();
+		
+
+	sec1 = 0;
+	sec0 = 0;
+
+	while(1) {
+		display_char(1, 3, sec1+0x30);
+		display_char(1, 4, sec0+0x30);
 	}
+}
 
-	void display_string(unsigned char x, unsigned char y, unsigned char *string)
-	{
-		unsigned char i;
-		
-		cmd_lr(0xB8+x);     // x address
-		if( y < 10 )
-		cmd_l(0x40+y*6+4); // y address
-		else
-		cmd_r(0x40+(y-10)*6); // y address
-		
-		while( *string != '\0' ){
-			if( y == 10 ){
-				cmd_r(0x40);
-			}
-			for(i = 0; i < 5; i++){
-				if(y < 10)
-				data_l(font[*string-0x20][i]);
-				else
-				data_r(font[*string-0x20][i]);
-			}
-			
-			if(y < 10)
-			data_l(0x00);
-			else
-			data_r(0x00);
-			
-			string++;
-			y++;
-		}
-	}
-
-	void lcd_clear()
-	{
-		unsigned char i, j, x, y;
-		x = 0xB8;                         /* X start address */
-		y = 0x40;                         /* Y start address */
-		for(i = 0; i < 8; i++)
-		{
-			cmd_lr(x);    cmd_lr(y);
-			for(j = 0; j < 64; j++)
-			data_lr(0x00);           /* clear CS1 and CS2 */
-			x++;
-		}
-		//HW: Clear ScreenBuffer as well
-	}
-
-	void fill_char_xy(unsigned char x_char, unsigned char y, unsigned char data)
-	{
-		cmd_lr( 0xB8 + x_char);         // X address
-		if(y <= 63)
-		{
-			cmd_l( 0x40 + y);      // CS1 Y address
-			data_l(data);
-		}
-		else
-		{
-			cmd_r( 0x40 + y - 64); // CS2 Y address
-			data_r(data);
-		}
-	}
-
-	void InitScreenBuffer( unsigned char ScreenBuffer[8][128] )
-	{
-		unsigned int i, j;
-		
-		for(i = 0 ; i < 8; i++){
-			for(j = 0 ; j < 128; j++){
-				ScreenBuffer[i][j] = 0;
-			}
-		}
-	}
-
-	void GLCD_Dot(unsigned char ScreenBuffer[8][128], unsigned char x, unsigned char y)
-	{
-		unsigned char x_char, xx, dot_data;
-		
-		//Error: out of screen
-		if((x >= 64) || (y >= 128)) return;
-		
-		x_char = x / 8;    // calculate x address (in terms of char)
-		xx = x % 8;        //                     (within a char)
-		
-		dot_data = (1<<xx);                  // Make a dot
-		ScreenBuffer[x_char][y] |= dot_data; // OR old data with dot data
-		// draw on GLCD screen
-		fill_char_xy(x_char, y, ScreenBuffer[x_char][y]);
-	}
-
-	void GLCD_Line(unsigned char ScreenBuffer[8][128], unsigned char x1,unsigned char y1,unsigned char x2,unsigned char y2)
-	{
-		int x, y;
-		if(y1 != y2)    // if y1 != y2, y is variable
-		{
-			if(y1 < y2)    //              x is function
-			{
-				for(y = y1; y <= y2; y++)
-				{
-					x = x1 + (y - y1)*(x2 - x1)/(y2 - y1);
-					GLCD_Dot(ScreenBuffer,x,y);
-				}
-			}
-			else
-			{
-				for(y = y1; y >= y2; y--)
-				{
-					x = x1 + (y - y1)*(x2 - x1)/(y2 - y1);
-					GLCD_Dot(ScreenBuffer,x,y);
-				}
-			}
-		}
-		else if(x1 != x2)   // if x1 != x2, x is variable
-		{
-			if(x1 < x2)    //              y is function
-			{
-				for(x = x1; x <= x2; x++)
-				{ y = y1 + (x - x1)*(y2 - y1)/(x2 - x1);
-					GLCD_Dot(ScreenBuffer,x,y);
-				}
-			}
-			else
-			{
-				for(x = x1; x >= x2; x--)
-				{
-					y = y1 + (x - x1)*(y2 - y1)/(x2 - x1);
-					GLCD_Dot(ScreenBuffer,x,y);
-				}
-			}
-		}
-		else                  // if x1 == x2 and y1 == y2,
-		{
-			GLCD_Dot(ScreenBuffer,x1,y1);  //             it is a dot
-		}
-	}
-
-	void GLCD_Rectangle(unsigned char ScreenBuffer[8][128],unsigned char x1,unsigned char y1,unsigned char x2,unsigned char y2)
-	{
-		//Fill your code here
-		
-	}
-
-	void GLCD_RectangleBlack(unsigned char ScreenBuffer[8][128],unsigned char x1,unsigned char y1,unsigned char x2,unsigned char y2) {
-		//HW
-	}
-
-	void GLCD_Circle(unsigned char ScreenBuffer[8][128],unsigned char x1,unsigned char y1,unsigned char r)
-	{
-		int x, y;
-		float s;
-		for(y = y1 - r*3/4; y <= y1 + r*3/4; y++) // draw with y variable
-		{
-			s = sqrt(r*r - (y-y1)*(y-y1)) + 0.5;
-			x = x1 + (unsigned char)s;
-			GLCD_Dot(ScreenBuffer,x,y);
-			x = x1 - (unsigned char)s;
-			GLCD_Dot(ScreenBuffer,x,y);
-		}
-		for(x = x1 - r*3/4; x <= x1 + r*3/4; x++) // draw with x variable
-		{
-			s = sqrt(r*r - (x-x1)*(x-x1)) + 0.5;
-			y = y1 + (unsigned char)s;
-			GLCD_Dot(ScreenBuffer,x,y);
-			y = y1 - (unsigned char)s;
-			GLCD_Dot(ScreenBuffer,x,y);
-		}
-	}
-
-	unsigned char sec0, sec1, sec2, sec3;
-
-	int main(void)
-	{
-		//Initialization
-		DDRA = 0xFF;
-		PORTB = 0xFF; DDRB = 0xFF;
-		DDRD = 0x0;
-		DDRE = 0xFF;
-		PORTG = 0x0; DDRG = 0xFF;
-
-		cmd_lr( 0x3F ); //DISP ON
-		cmd_lr( 0xB8 );
-		cmd_lr( 0x40 );
-		
-		lcd_clear();
-		display_string(0, 0, "Clock Interrupt Test");
-		
-		TCNT1H = 0xB;
-		TCNT1L = 0xDC;
-		TCCR1A = 0x0;
-		TCCR1B = 0x04;
-		
-		TIMSK = (1<<TOIE1);
-		sei();
-		
-		sec3 = 0;
-		sec2 = 0;
-		sec1 = 0;
+ISR(TIMER1_OVF_vect){
+	sec0++;
+	if(sec0 == 10){
 		sec0 = 0;
-
-		while(1){
-			display_char(1, 0, sec3+0x30);
-			display_char(1, 1, sec2+0x30);
-			display_char(1, 2, 0x3a);
-			display_char(1, 3, sec1+0x30);
-			display_char(1, 4, sec0+0x30);
+		sec1++;
+		if(sec1 == 6){
+			sec1 = 0;
 		}
 	}
-
-	ISR(TIMER1_OVF_vect){
-		sec0++;
-		if(sec0 == 10){
-			sec0 = 0;
-			sec1++;
-			if(sec1 == 6){
-				sec1 = 0;
-				sec2++;
-				if(sec2 == 10) {
-					sec2 = 0;
-					sec3++;
-					if (sec3 == 6) {
-						sec3 = 0;
-					}
-				}
-			}
-		}
-	}
+}
